feat(blockchain)

+++ src/modules/users/service.ts
@@ -10,16 +10,17 @@ import { hashPassword, verifyPassword } from "../../shared/utils/password";
 import { getTokenPayload, generateTokens, verifyToken } from "../../shared/utils/jwt";
 import type { RegisterInput, LoginInput } from "./validation";
 import { translateCity } from "../../shared/utils/translations";
-import { emitEvent, USER_EVENTS } from "../../shared/event-bus";
-import { asc, eq, sql } from "drizzle-orm";
+import { asc, count, eq, sql } from "drizzle-orm";
 import { cities, citiesCN, citiesRU, City, Club, clubs, NewUser, User, users } from "./schema";
 import { userRepository } from "./index";
 import { isAdmin, removeEmptyFields } from "../../shared/utils/helpers"
+import { emitEvent, USER_EVENTS } from "../../shared/event-bus";
 
 const getUser = async (lang: string, user: User & { city: City, club: Club }) => ({
   id: user.id,
   email: user.email,
   username: user.username,
+  image: user.image,
   gender: user.gender,
   city: {
     ...user.city,
@@ -27,7 +28,8 @@ const getUser = async (lang: string, user: User & { city: City, club: Club }) =>
   },
   moderatorTournamentsIds: user.moderatorTournamentsIds,
   club: user.club,
-  isAdmin: user.isAdmin,
+  isAdmin: user.isAdmin || user.email === Bun.env.ADMIN_EMAIL,
+  blockchainId: user.blockchainId,
   totalMatches: user.totalMatches,
   createdAt: user.createdAt.toISOString()
 })
@@ -138,19 +140,22 @@ export class AuthService {
 }
 
 export class UserService {
-  async getAll(lang?: string) {
+  async getAll(page: number, pageSize=10, lang?: string) {
     const usersArr = await db.query.users.findMany({
       with: {
         city: true,
         club: true
       },
+      limit: pageSize,
+      offset: (page - 1) * pageSize,
       orderBy: asc(users.createdAt)
     })
     const result: User[] = new Array(usersArr.length)
     for (let i = 0; i < result.length; i++) {
       result[i] = await getUser(lang||"en", usersArr[i])
     }
-    return result
+    const usersCount = (await db.select({ count: count() }).from(users))[0].count;
+    return { users: result, usersCount }
   }
 
   async getByClubId(clubId: number) {
@@ -182,12 +187,15 @@ export class UserService {
       passwordHash = await hashPassword(password);
     }
 
-    if (await isAdmin(actorId) && adminIs) {
+    if (await isAdmin(actorId, true) && adminIs) {
       isAdminVar = adminIs
     }
     const pureData = removeEmptyFields(other)
     await db.update(users)
-      .set(passwordHash ? {...pureData, isAdmin: isAdminVar, passwordHash, updatedAt: new Date()} : {...pureData, isAdmin: isAdminVar, updatedAt: new Date()})
+      .set(passwordHash ?
+        {...pureData, isAdmin: isAdminVar, passwordHash, syncedToBlockchain: false, updatedAt: new Date()}
+        :
+        {...pureData, isAdmin: isAdminVar, syncedToBlockchain: false, updatedAt: new Date()})
       .where(eq(users.id, userId))
 
     const user = await db.query.users.findFirst({
@@ -197,7 +205,7 @@ export class UserService {
         club: true
       }
     })
-    return await getUser(lang, user)
+    return await getUser(lang, user!)
   }
 
   async updateUserTournamentModerator(tournamentId: number, userId: string) {
@@ -218,19 +226,51 @@ export class UserService {
     await db.update(users)
       .set({
         moderatorTournamentsIds: newIds,
-        updatedAt: new Date()
+        updatedAt: new Date(),
+        syncedToBlockchain: false
       })
       .where(eq(users.id, userId));
     }
 
-    async updateUserMatches(matchesPlayed: number, userId: string) {
-      await db.update(users)
-        .set({
-          totalMatches: sql`${users.totalMatches} + ${matchesPlayed}`,
-          updatedAt: new Date()
-        })
-        .where(eq(users.id, userId));
+  async deleteTournamenModerator(tournamentId: number, userId: string) {
+    const user = await db.query.users.findFirst({
+      where: eq(users.id, userId),
+      columns: {
+        moderatorTournamentsIds: true
+      }
+    });
+
+    if (!user || !user.moderatorTournamentsIds) return
+
+    const newTournamentsIds = user.moderatorTournamentsIds.filter(id=>id !== tournamentId)
+
+    await db.update(users)
+      .set({
+        moderatorTournamentsIds: newTournamentsIds,
+        updatedAt: new Date(),
+        syncedToBlockchain: false
+      })
+      .where(eq(users.id, userId));
+  }
+
+  async updateUserMatches(matchesPlayed: number, userId: string) {
+    await db.update(users)
+      .set({
+        totalMatches: sql`${users.totalMatches} + ${matchesPlayed}`,
+        updatedAt: new Date(),
+        syncedToBlockchain: false
+      })
+      .where(eq(users.id, userId));
+  }
+
+  async delete(id: string, actorId: string) {
+    if (await isAdmin(actorId) || actorId === id) {
+      await db.delete(users).where(eq(users.id, id))
+      return { success: true }
+    } else {
+      throw new Error("You don't admin or this user")
     }
+  }
 }
 
 export class ClubService {
@@ -267,7 +307,7 @@ export class ClubService {
 
 export class CityService {
   async getAll(lang: string) {
-    const citiesArr = await db.select().from(cities);
+    const citiesArr = await db.select().from(cities).orderBy(asc(cities.id));
     if (lang === "en") {
       return citiesArr;
     } else {
@@ -289,8 +329,9 @@ export class CityService {
   }
 
   async create(title: string) {
+    const translations = await userRepository.translateCity(title)
     const isCity = await db.query.cities.findFirst({
-      where: eq(cities.title, title),
+      where: eq(cities.title, translations.en),
     });
     if (isCity) {
       throw new Error("City is exist");
@@ -300,33 +341,30 @@ export class CityService {
       .insert(cities)
       .values({ title })
       .returning();
+    await db
+      .insert(citiesRU)
+      .values({ title: translations.ru, id: city.id })
+    await db
+          .insert(citiesCN)
+          .values({ title: translations.zh, id : city.id })
     return city;
   }
 
-  async delete(id: number) {
+  async update(id: number, title: string, lang: string) {
+    if (lang === "en") {
+      await db.update(cities).set({ title }).where(eq(cities.id, id))
+    } else if (lang === "ru") {
+      await db.update(citiesRU).set({ title }).where(eq(citiesRU.id, id))
+    } else if (lang === "zh") {
+      await db.update(citiesCN).set({ title }).where(eq(citiesCN.id, id))
+    }
+    return { success: true };
+  }
+
+  async delete(id: number, actorId: string) {
+    if (!await isAdmin(actorId))
+      throw new Error("You don't admin")
     await db.delete(cities).where(eq(cities.id, id));
-    return { success: true };
-  }
-
-  async createRU(title: string) {
-    await db
-          .insert(citiesRU)
-          .values({ title })
-  }
-
-  async createCN(title: string) {
-    await db
-          .insert(citiesCN)
-          .values({ title })
-  }
-
-  async deleteAllRU() {
-    await db.delete(citiesRU);
-    return { success: true };
-  }
-
-  async deleteAllCN() {
-    await db.delete(citiesCN);
     return { success: true };
   }
 }
